tests/wpt/web-platform-tests/navigation-api/ordering-and-transition/anchor-download.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);

  let navigate_called = false;
  i.contentWindow.navigation.onnavigate = () => navigate_called = true;
  navigation.onnavigate = t.unreached_func("navigate must not fire");

  let a = i.contentDocument.createElement("a");
  a.href = "?1";
  a.download = "";
  i.contentDocument.body.appendChild(a);
  a.click();

  i.contentWindow.navigation.onnavigatesuccess = t.unreached_func("navigatesuccess must not fire");
  i.contentWindow.navigation.onnavigateerror = t.unreached_func("navigateerror must not fire");
  await new Promise(resolve => t.step_timeout(resolve, 20));
  assert_true(navigate_called);
}, "<a download> fires navigate, but not navigatesuccess or navigateerror when not intercepted by intercept()");
</script>
